pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.isort]
profile = "black"

[tool.poetry]
authors = ["Guilhem Saurel <guilhem.saurel@laas.fr>"]
description = "Post a message to a matrix room with a simple HTTP POST"
homepage = "https://github.com/nim65s/matrix-webhook"
license = "BSD-2-Clause"
name = "matrix-webhook"
readme = "README.md"
repository = "https://github.com/nim65s/matrix-webhook.git"
version = "3.7.0"

[tool.poetry.dependencies]
Markdown = "^3.3.4"
matrix-nio = ">=0.18.3,<0.21.0"
python = "^3.8"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
coverage = ">=6.4.4,<8.0.0"
flake8 = "^5.0.4"
httpx = "^0.23.0"
isort = "^5.12.0"
pydocstyle = "^6.1.1"
safety = {allow-prereleases = true, version = "^2.4.0b1"}

[tool.poetry.scripts]
matrix-webhook = "matrix_webhook.__main__:main"

[tool.poetry.urls]
"changelog" = "https://github.com/nim65s/matrix-webhook/blob/master/CHANGELOG.md"

[tool.ruff]
extend-ignore = ["D203", "D213"]
extend-select = ["A", "B", "C", "COM", "D", "EM", "EXE", "G", "N", "PTH", "RET", "RUF", "UP", "W", "YTT"]
target-version = "py38"

[tool.tomlsort]
all = true
